Automat and Bottler Controllers are logged

+++ src/main/java/com/sirius/web/controller/AutomatController.java
@@ -4,6 +4,8 @@ import com.sirius.web.model.Automat;
 import com.sirius.web.model.Bottle;
 import com.sirius.web.service.AutomatService;
 import com.sirius.web.service.BottleService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -16,6 +18,7 @@ import java.util.List;
 @RequestMapping("/rest/automats")
 public class AutomatController implements Serializable {
 
+    private static Logger logger = LoggerFactory.getLogger(AutomatController.class);
     private final AutomatService automatService;
     private final BottleService bottleService;
 
@@ -30,9 +33,10 @@ public class AutomatController implements Serializable {
         final List<Automat> automats = automatService.getAllAutomats();
 
         if(automats == null) {
+            logger.error("Kayitli otomatlar listenemedi");
             return new ResponseEntity<List<Automat>>(HttpStatus.NOT_FOUND);
         }
-
+        logger.info("Kayitli otomat listesi: " + automats);
         return new ResponseEntity<List<Automat>>(automats, HttpStatus.OK);
     }
 
@@ -41,8 +45,10 @@ public class AutomatController implements Serializable {
 
         if (automat.getId() != null && automat.getId().length() != 0 && !automatService.exists(automat.getId())) {
             final Automat dbautomat = automatService.createAutomat(automat);
+            logger.info("Yeni otomat olusturuldu. Otomat bilgileri: " + automat);
             return new ResponseEntity<Automat>(dbautomat, HttpStatus.CREATED);
         } else {
+            logger.error("Yeni otomat olusturalamadi");
             return new ResponseEntity<Automat>(HttpStatus.BAD_REQUEST);
         }
 
@@ -50,13 +56,12 @@ public class AutomatController implements Serializable {
 
     @GetMapping("/{id}")
     public ResponseEntity<Automat> getAutomat(@PathVariable String id) {
-        try {
-            final Automat dbautomat = automatService.findAutomatById(id);
-            if (dbautomat == null) {
-                return new ResponseEntity<Automat>(HttpStatus.NOT_FOUND);
-            }
-            return new ResponseEntity<Automat>(dbautomat, HttpStatus.OK);
-        } catch (Exception e) {
+        final Automat dbAutomat = automatService.findAutomatById(id);
+        if (dbAutomat != null) {
+            logger.info(id + " ID numarali otomatin bilgileri getirildi");
+            return new ResponseEntity<Automat>(dbAutomat, HttpStatus.OK);
+        } else {
+            logger.error(id + " ID numarali otomat bulunamadi");
             return new ResponseEntity<Automat>(HttpStatus.NOT_FOUND);
         }
     }
@@ -67,28 +72,31 @@ public class AutomatController implements Serializable {
         final boolean isExist = automatService.exists(automat.getId());
         if (isExist) {
             automatService.updateAutomat(automat);
+            logger.info(automat.getId() + " ID numarali otomatin bilgileri guncellenmistir");
             return new ResponseEntity<Automat>(automat, HttpStatus.OK);
         } else {
+            logger.error(automat.getId() + " ID numarali otomat bulunamadi");
             return new ResponseEntity<Automat>(HttpStatus.BAD_REQUEST);
         }
     }
 
     @DeleteMapping("deleteAutomat")
     public ResponseEntity deleteAutomat(@RequestParam String id){
-        try {
-            boolean ok = automatService.deleteAutomat(id);
-            return new ResponseEntity(ok, HttpStatus.OK);
-        } catch (RuntimeException e){
-            return new ResponseEntity(false, HttpStatus.UNAUTHORIZED);
-        } catch (Exception e){
+        if(automatService.exists(id)) {
+            automatService.deleteAutomat(id);
+            logger.info(id + " ID numarali otomat silindi");
+            return new ResponseEntity(true, HttpStatus.OK);
+        } else {
+            logger.error(id + " ID numarali otomat silinemedi");
             return new ResponseEntity(false, HttpStatus.BAD_REQUEST);
         }
     }
 
     @DeleteMapping("deleteAll")
     public ResponseEntity deleteAll() {
-        boolean ok = automatService.deleteAll();
-        return new ResponseEntity(ok, HttpStatus.NO_CONTENT);
+        automatService.deleteAll();
+        logger.info("Tüm otomatlar silindi");
+        return new ResponseEntity(true, HttpStatus.NO_CONTENT);
     }
 
     @PutMapping("changeCapacity/{id}/{barcode}")
@@ -98,6 +106,7 @@ public class AutomatController implements Serializable {
         boolean isDbBottle = bottleService.exists(barcode);
 
         if (!isDbAutomat || !isDbBottle) {
+            logger.error(id + " veya " + barcode + " yanlis");
             return new ResponseEntity(false, HttpStatus.BAD_REQUEST);
         }
         else {
@@ -105,6 +114,7 @@ public class AutomatController implements Serializable {
             Bottle getBottleFromDb = bottleService.findBottleByBarcode(barcode);
 
             if(!getAutomatFromDb.isActive()) {
+                logger.error(id + " ID numarali otomat aktif degil");
                 return new ResponseEntity(false, HttpStatus.BAD_REQUEST);
             }
 
@@ -114,11 +124,13 @@ public class AutomatController implements Serializable {
 
             if(getAutomatFromDb.getCapacity() <= 100.00) {
                 automatService.updateAutomat(getAutomatFromDb);
+                logger.info(id + " ID numarali otomatin yeni kapasitesi: " + getAutomatFromDb.getCapacity());
                 return new ResponseEntity(true, HttpStatus.OK);
             }
             else {
                 getAutomatFromDb.setActive(false);
                 automatService.updateAutomat(getAutomatFromDb);
+                logger.error(id + " ID numarali otomatin kapasitesi doludur");
                 return new ResponseEntity(false, HttpStatus.BAD_REQUEST);
             }
         }
